+++ test/tap-parallel-not-ok/todo/update-todo.test.js
@@ -1,6 +1,9 @@
 import tap from 'tap';
-import { build } from '../../src/app.js';
+import { build } from '../../../src/app.js';
 import 'must/register.js';
+import Chance from 'chance';
+
+const chance = new Chance();
 
 tap.mochaGlobals();
 
@@ -10,7 +13,62 @@ describe('Update a todo should work', async () => {
   let app;
 
   before(async () => {
-    app = await build();
+    app = await build({
+      forceCloseConnections: true
+    });
+  });
+
+  const newUser = {
+    username: chance.email({ domain: 'example.com' }),
+    password: chance.string({ length: 12 }),
+    firstName: chance.first(),
+    lastName: chance.last()
+  };
+
+  let cookie = '';
+
+  it('Should return the user that was created a new user', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/register`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(newUser)
+    });
+
+    // this checks if HTTP status code is equal to 200
+    response.statusCode.must.be.equal(200);
+
+    const result = await response.json();
+
+    // expect that id exists
+    result.username.must.be.equal(newUser.username);
+    result.firstName.must.be.equal(newUser.firstName);
+    result.lastName.must.be.equal(newUser.lastName);
+
+    // expect createdDate and updatedDate is not null
+    result.createdDate.must.not.be.null();
+    result.updatedDate.must.not.be.null();
+  });
+
+  it('Login should work', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/login`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        username: newUser.username,
+        password: newUser.password
+      })
+    });
+
+    // this checks if HTTP status code is equal to 200
+    response.statusCode.must.be.equal(200);
+
+    cookie = response.headers['set-cookie'];
   });
 
   it('Should update the object given an ID', async () => {
@@ -29,7 +87,8 @@ describe('Update a todo should work', async () => {
       method: 'POST',
       url: `${prefix}/todo`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newTodo)
     });
@@ -40,7 +99,8 @@ describe('Update a todo should work', async () => {
       method: 'PUT',
       url: `${prefix}/todo/${id}`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newerTodo)
     });
@@ -76,7 +136,8 @@ describe('Update a todo should work', async () => {
       method: 'POST',
       url: `${prefix}/todo`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newTodo)
     });
@@ -87,7 +148,8 @@ describe('Update a todo should work', async () => {
       method: 'PUT',
       url: `${prefix}/todo/${id}`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newerTodo)
     });
@@ -123,7 +185,8 @@ describe('Update a todo should work', async () => {
       method: 'POST',
       url: `${prefix}/todo`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newTodo)
     });
@@ -134,7 +197,8 @@ describe('Update a todo should work', async () => {
       method: 'PUT',
       url: `${prefix}/todo/${id}`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newerTodo)
     });
@@ -154,5 +218,10 @@ describe('Update a todo should work', async () => {
     // expect createdDate and updatedDate is not null
     result.createdDate.must.equal(createdDate);
     result.updatedDate.must.above(updatedDate);
+  });
+
+  
+  after(async () => {
+    await app.close();
   });
 });
