+++ test/tap-parallel-not-ok/user/login.test.js
@@ -1,5 +1,5 @@
 import tap from 'tap';
-import { build } from '../../src/app.js';
+import { build } from '../../../src/app.js';
 import 'must/register.js';
 import Chance from 'chance';
 
@@ -9,11 +9,13 @@ tap.mochaGlobals();
 
 const prefix = '/api';
 
-describe('Logging out a user should work', async () => {
+describe('Logging in a user should work', async () => {
   let app;
 
   before(async () => {
-    app = await build();
+    app = await build({
+      forceCloseConnections: true
+    });
   });
 
   const newUser = {
@@ -22,8 +24,6 @@ describe('Logging out a user should work', async () => {
     firstName: chance.first(),
     lastName: chance.last()
   };
-
-  let cookie = '';
 
   it('Should return the user that was created a new user', async () => {
     const response = await app.inject({
@@ -65,34 +65,43 @@ describe('Logging out a user should work', async () => {
 
     // this checks if HTTP status code is equal to 200
     response.statusCode.must.be.equal(200);
-
-    cookie = response.headers['set-cookie'];
   });
 
-  it('Logout should work', async () => {
+  it('Login should return an error if username doesn\'t exist', async () => {
     const response = await app.inject({
-      method: 'GET',
-      url: `${prefix}/logout`,
-      headers: {
-        'Content-Type': 'application/json',
-        cookie
-      }
-    });
-
-    // this checks if HTTP status code is equal to 401
-    response.statusCode.must.be.equal(200);
-  });
-
-  it('Logout should return an error without a cookie', async () => {
-    const response = await app.inject({
-      method: 'GET',
-      url: `${prefix}/logout`,
+      method: 'POST',
+      url: `${prefix}/login`,
       headers: {
         'Content-Type': 'application/json'
-      }
+      },
+      body: JSON.stringify({
+        username: 'test',
+        password: 'password'
+      })
     });
 
-    // this checks if HTTP status code is equal to 401
+    // this checks if HTTP status code is equal to 200
     response.statusCode.must.be.equal(401);
+  });
+
+  it('Login should return an error if password is incorrect', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/login`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        username: newUser.username,
+        password: 'password'
+      })
+    });
+
+    // this checks if HTTP status code is equal to 200
+    response.statusCode.must.be.equal(401);
+  });
+  
+  after(async () => {
+    await app.close();
   });
 });
