+++ test/tap-parallel-not-ok/user/logout.test.js
@@ -1,5 +1,5 @@
 import tap from 'tap';
-import { build } from '../../src/app.js';
+import { build } from '../../../src/app.js';
 import 'must/register.js';
 import Chance from 'chance';
 
@@ -9,11 +9,13 @@ tap.mochaGlobals();
 
 const prefix = '/api';
 
-describe('Logging in a user should work', async () => {
+describe('Logging out a user should work', async () => {
   let app;
 
   before(async () => {
-    app = await build();
+    app = await build({
+      forceCloseConnections: true
+    });
   });
 
   const newUser = {
@@ -22,6 +24,8 @@ describe('Logging in a user should work', async () => {
     firstName: chance.first(),
     lastName: chance.last()
   };
+
+  let cookie = '';
 
   it('Should return the user that was created a new user', async () => {
     const response = await app.inject({
@@ -63,39 +67,38 @@ describe('Logging in a user should work', async () => {
 
     // this checks if HTTP status code is equal to 200
     response.statusCode.must.be.equal(200);
+
+    cookie = response.headers['set-cookie'];
   });
 
-  it('Login should return an error if username doesn\'t exist', async () => {
+  it('Logout should work', async () => {
     const response = await app.inject({
-      method: 'POST',
-      url: `${prefix}/login`,
+      method: 'GET',
+      url: `${prefix}/logout`,
       headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        username: 'test',
-        password: 'password'
-      })
+        'Content-Type': 'application/json',
+        cookie
+      }
     });
 
-    // this checks if HTTP status code is equal to 200
-    response.statusCode.must.be.equal(401);
+    // this checks if HTTP status code is equal to 401
+    response.statusCode.must.be.equal(200);
   });
 
-  it('Login should return an error if password is incorrect', async () => {
+  it('Logout should return an error without a cookie', async () => {
     const response = await app.inject({
-      method: 'POST',
-      url: `${prefix}/login`,
+      method: 'GET',
+      url: `${prefix}/logout`,
       headers: {
         'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        username: newUser.username,
-        password: 'password'
-      })
+      }
     });
 
-    // this checks if HTTP status code is equal to 200
+    // this checks if HTTP status code is equal to 401
     response.statusCode.must.be.equal(401);
+  });
+  
+  after(async () => {
+    await app.close();
   });
 });
